requirements.txt
# Core Data Libraries
pandas==2.1.4
numpy==1.24.4

# Visualization for EDA (optional but useful)
matplotlib==3.8.2
seaborn==0.13.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# FastAPI API Server
fastapi==0.110.1
uvicorn[standard]==0.27.1

# Input validation
pydantic==2.6.4

# PostgreSQL DB logging
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29

# Optional: AI API Integration (OpenAI)
openai==1.3.7

# Optional: Local model explainability (don't use in production Docker)
shap==0.44.1

#used for making http requests 
httpx==0.27.0

python-dotenv
